feat : 메시지 대기열 실시간 확인
