--- src/main/java/com/dnd/reevserver/global/common/service/MessageService.java
@@ -1,25 +1,0 @@
-package com.dnd.reevserver.global.common.service;
-
-import com.amazonaws.services.sqs.AmazonSQSAsync;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cloud.aws.messaging.core.QueueMessagingTemplate;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.support.MessageBuilder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MessageService {
-    private final QueueMessagingTemplate queueMessagingTemplate;
-
-    @Value("${cloud.aws.sqs.queue-name}")
-    private String queueName;
-
-    public MessageService(AmazonSQSAsync amazonSQS){
-        this.queueMessagingTemplate = new QueueMessagingTemplate(amazonSQS);
-    }
-
-    public void sendMessage(String message){
-        Message<String> newMessage = MessageBuilder.withPayload(message).build();
-        queueMessagingTemplate.send(queueName, newMessage);
-    }
-}
